Updated makercheck approval logic

+++ backend_rest/makerchecker/executor.py
@@ -1,20 +1,30 @@
+from backend_rest.core import models
 from sales import models as m_sales
+from . import models as mc_models
 
 
-def delete_sale_docs(task):
+def delete_sale_docs(task, status):
     sale_id = int(task.reference)
-    m_sales.Document.objects.filter(sale_id=sale_id).delete()
-    m_sales.Sale.objects.filter(pk=sale_id).update(agent=None, task=None)
     sale = m_sales.Sale.objects.get(pk=sale_id)
-    return f'Documents deleted successfully from sale: {sale.sales_order}'
+    if status == mc_models.STATUS_APPROVED:
+        m_sales.Document.objects.filter(sale_id=sale_id).delete()
+        m_sales.Sale.objects.filter(pk=sale_id).update(agent=None, task=None)
+        return f'Documents deleted successfully from sale: {sale.sales_order}'
+    else:
+        return f'Documents deleted successfully rejected for sale: {sale.sales_order}'
 
 
-def waive_missing_c2(task):
+def waive_missing_c2(task, status):
     sale_id = int(task.reference)
-    m_sales.Document.objects.filter(sale_id=sale_id).delete()
-    m_sales.Sale.objects.filter(pk=sale_id).update(agent=None, task=None)
     sale = m_sales.Sale.objects.get(pk=sale_id)
-    return f'Documents deleted successfully from sale: {sale.sales_order}'
+    if status == mc_models.STATUS_APPROVED:
+        m_sales.Document.objects.filter(sale_id=sale_id).update(status=1)
+        m_sales.Sale.objects.filter(pk=sale_id).update(task=None)
+        return f'Documents with missing C2 successfully approved for sale: {sale.sales_order}'
+    else:
+        m_sales.Document.objects.filter(sale_id=sale_id).delete()
+        m_sales.Sale.objects.filter(pk=sale_id).update(agent=None, task=None)
+        return f'Documents with missing C2 successfully rejected for sale: {sale.sales_order}'
 
 
 EXECUTOR_CHOICES = [
